Listeners for GUI and rename
git-svn-id: http://uuh116.googlecode.com/svn/Nebulabrot@167 333f4dc9-c705-684f-d081-88386eedeb15

+++ src/net/lab0/nebula/data/SynchronizedCounter.java
@@ -21,7 +21,7 @@ public class SynchronizedCounter
         ++count;
     }
     
-    public synchronized void increment(int quantity)
+    public synchronized void increment(long quantity)
     {
         count += quantity;
     }
@@ -31,7 +31,7 @@ public class SynchronizedCounter
         --count;
     }
     
-    public synchronized void decrement(int quantity)
+    public synchronized void decrement(long quantity)
     {
         count -= quantity;
     }
